pyproject.toml
[project]
name = "arachne"
version = "0.1.0"
description = "Table classification system for insurance tender documents"
readme = "README.md"
requires-python = "==3.13"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.5.0",
    "psycopg2-binary>=2.9.9",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "joblib>=1.4.0",
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "scipy>=1.17.1",
]

[project.optional-dependencies]
transformers = [
    "torch>=2.4.0",
    "transformers>=4.44.0",
    "accelerate>=0.33.0",
    "sentencepiece>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
